Test: Link simple.rst
We add a `.end` directive to simple.a65 so that we demonstrate how to
set the address in the entry point record in the output file. (It was
defaulting to 0000.) This also shows that `.end` doesn't actually end
assembly.

Also fix a bug in our list file post-processing that was causing the
.rst file generated by the linker to be wrong. (diff showed that it
was editing the wrong lines from the .lst file if we replaced FF with
LF before we linked.)