Introduce pytest framework
We don't actually test anything yet (though we do show that a test
runs); this is just the framework.

I had hoped to name the test file for `simple.a65`, `simple.pt`, as
I've done before, but I seem to have issues with getting that to work.
So we go with the uglier default of `simple_test.py` for now and I'll
work out how to fix this later.

Using `-o 'pytest_files=*.pt` on the command line before giving the
directory for discovery also threw an odd error about not being able
to find configuration key `testpaths` in pytest 5.0 and 3.0 (but
not 3.5).