Test: Pass all command-line args on to pytest
This allows the developer to set verbosity and other options and
choose which tests to discover and run.

It turns out we don't need special handling with pytest to remove the
`-q` we supply by default; the `-v` supplied by the user will
counteract it.

Now that the user can provide dirs/files for discovery, we need not to
provide our default ones if he does that. Pytest directly supports
this beahviour with the `testpaths` configuration variable.