src/simple: Test addxy and fix the bug in it
The bug was that it was not storing the result in xybuf as well as
returning it in the A register.

A few framework extensions were hacked in to be able to do this test:
a count parameter for step(), strAt() to get data out of memory as a
string, and byteAt() and wordAt() calling through to the underlying MPU.

This is now exposing a lot of awkwardness in our interfaces
(particularly assertregs(), for which I've thought of a solution), so
it's time to start cleaning this up and improving the framework.