Test: Remove headings from assembly listings
Once my listing became more than a page long, the headings and page
breaks were landing in the middle of routines I was trying to read.

This doesn't suppress the formfeeds/headings (now "Page 1" and "Page
2) before the symbol and area tables at the end of the listing. Those
could go in a separate file, but that would be more inconvenient than
the headings.

In fact, now that there's only those two headings at the end, dropping
the post-processing to remove the formfeeds is probably reasonable
since just the two formfeed characters won't stick out so much.