pylib/tmpu: Replace assertregs() with a Regs class and mpu.regs value
This brings the actual assertion back from the test support framework
into the test module itself, so that the assertion is properly
instrumented by the pytest compiler. It also gives us a nice repr()
that makes it easy to see the values of registers and flags. These
values may be unset (`None`) in Regs objects you create, in which
case that signals that those values should not be compared.