pylib/tmpu: Rename TMPU → Machine; use M as local variable convention
TMPU was a somewhat mysterious class name, and it emulates the entire
system, not just the CPU or MPU, so call it `Machine` and we'll change
the module name later on to make it more clear what kind of "Machine"
this is. (I also thought of using `System`, but that was a bit too vague,
and also the `S` abbreviation of it collides with `SymTab`--see below.)

I also extract a fixture to generate the machine instance for tests
and change my local variable convention to use `M` as the name of
that. Yes, I know that's normally used for "constants," but it matches
the `R` abbreviation for construction a `Registers` instance. So the
convention seems to be becoming, here, "use a single capital letter
for the test machine/system items we're manipulating." This keeps them
short (good because they're used a _lot_) and easily distinguished
from all the other bits of code.

The module renaming for `tmpu` will come soon.