Rename pylib/tmpu → lib/testmc/m6502
We move our library directory from pylib/ to just lib/ because that's
a bit shorter and makes it easier to use command-line completion.
There's no interference with our "real" source code and tests because
that's always all under src/. And that directory will eventually not
be in repos like this because it will be a separate Python package.

We rename the `tmpu` module to `testmc.m6502`, `testmc` meaning "test
microcomputer (or microcontroller) system." The particular CPU gets
extracted out to its own module under testmc because we will
eventually be supporting MC6809 and Z80 as well. (The only reason for
the `m` prefix on `6502` is because module names must start with a
letter in order to be able to use them in import statements.)

I've checked with PyPI and `testmc` doesn't collide with any package
names there. (`testmachine` was another option I'd considered, but
someone already has a package with that name.)