src/reloctest*: Add relocatable source file with ident and xfailed test
This is a relocatable object, rather than an absolute object as with
simple.a65. The Test script has been modified to build it, setting the
start of the segment to 0x400. (Actually, the ident string should
probably be in a separate data segment; we'll figure out how to do
that later.)

This fails because the symbol table in the linker listing still says
that the object is at 0000 even though it's been relocated to 0400 (as
shown in the listing itself). We need to fix our symbol table loader
to handle this before this will pass.