src/simple: `bsread` routine to read in bytes from memory
This is the 6502 version of a BytesIO(), allowing our tests to
store data in memory and have this read them as a stream, used
with routines designed to read from an IO device.

We also add a new area, ZP, for the zero page, allowing us to
allocate addresses there inline with the routines that use them.