testmc.m6502.call(): Add list of instructions on which to abort
The default list of abort instructions is [BRK], since in most
circumstances you wouldn't want to be executing this (and you may be
in an unset area of memory on the 6502 emulator).

We add a new Abort exception for this, and make it and the Timeout
exception include the registers and the next instruction in the
message.

Also improve some docstrings.