testmc.m6502: "instrs" → "opcodes" and add tracing to step()
"Instr" was a bit obscure and abbreviated, and I tend to think of
instructions as being assembler-level operations that have
multiple translations (e.g., LDX = any of $A2, $A6, $B6, $AE,
$BE) anyway, so we now use the term "opcode" everywhere we're
talking about the binary values.

step() and all the functions that call it now take a `trace`
parameter that will cause step() to print the current machine
state and next opcode before executing the step. This is False by
default so that failing tests don't produce a lot of noisy
output.

Also fix a bug in call() where it was always passing MAXOPS to
stepto() rather than the maxops parameter it had been given, and
fix its docstring to distinguish `Abort` from `Timeout`.