testmc/asxxxx: Implement relocatable area patchups for symbol tables
Symbol tables now have the correct relocated addresses when a single
relocatable module is linked with user-specified addresses for one or
more areas (-b option to aslink).

This works only for a single module. If multiple modules are linked
together it's probably still correct for the one module named the same
as the .bin file, but there will be no symbols for the other modules
becuase we read only one .sym/.lst/.rst file. This is probably not an
issue for unit tests because it seems reasonable to link each module
into a separate .bin for unit testing, though if one module depends on
another (that's presumably already been tested by itself) that should
still work.

We commit ASxxxx input and output files in testmc/testfiles/ (with
a Makefile to build them) in testmc/testfile/ so that we're reading
actual ASxxxx output for the routines that read those.

We deliberately specify the segment as lower-case `_code` on the
command line so that we also check that we're not treating area names
as case-sensitive. But we still do treat them as case-sensitive
in the SymTab API, which should be fixed.

Things still to be done:
- src/reloctest* should probably be moved into the tests for
  lib/testmc/ by generating a .bin file in testmc/testfiles/ and
  reading that.
- Confirm we have a 'Hexadecimal [16-Bits]' header in the
  .sym/.lst/.rst/.map files and generate an error if we're using
  another radix.
- Check symbols names read from both narrow and wide format output
  files and warn if any of them are at the maximum length that can be
  represented in those files: 14 for narrow, 55 for wide, both less
  than the 79 character maximum ASxxxx symbol length.
- Change SymTab API not to treat area names as case-sensitive.
- Perhaps find a way to deal with finding test files if we're not
  running from the repo's top-level directory.