Test,src/simple-asl: Start assembling with Macro Assembler AS
The simple* files now have and will keep versions for each toolchain
for testing and experimentation. For ASxxxx we use the -asx suffix for
brevity; for Macro Assembler AS we use the -asl suffix since that's
the name of the actual assembler command (and -as is too vague).

The new -asl code copies a simple routine from the -asx code so we
can test assembly and see the output files. There are no tests
because the test framework does not yet support loading AS binary
or symbol files.