Test asbuild(): Run asl in object dir and deal with various issues
AS has some behaviour related to input and output file locations that
we need to work around, so we extract a function and handle all this
there. See the comments in the source for details. This also now
disables page headings where possible and adds the -s option for a
hierarchical listing of section names in the .lst file.

We're intending in the long run to have at least two assembler source
directories: src/ for individual modules and sys/ for the "top-level"
files that build a particular configuration of an application by
setting variables and including the appropriate modules.

So currently our convention is that all include files in the source
will use a path relative to $basedir; this seems the most likely to
avoid confusion and conflicts, at the moment. This may need to be
reviewed later.