lib/testmc/testfiles/asl/: AS output files for upcoming testmc.asl module
We use equates to generate all types of symbol values (Int, Float and
String) and pay particular attention to generating "unusual" string
values that should generate backslash-interpolated values in the
output, including `\000` bytes. The output was built with a version of
AS that includes the fix for the bug (present from 2008 up to at least
asl-current-142-bld151) that was generating spaces rather than `\032`
in string values.

We generate address-valued symbols in pretty much every scope type we
can think of. These turn in to special symbol names in the symbol
table, as well as generating "Info for Section" sections for local
scopes.

We generate symbols only in the NOTHING and CODE segments (which we
call "sections" in testmc), but because we already handle two sections
we should be fine for any arbitrary number of them.