testmc.memimage: Iterating over MemImage always produces (addr,data) tuples
`for` had always produced (addr,data) tuples to this point because the
MemRecords we used were always tuples containing just those two
values, in that order.

However in the future we wish to keep further information about each
record, such as the segement to which it belongs (CODE, DATA, etc.)
and so on. So we now add an __iter__() that produces a generator
extracting just the `addr` and `data` attributes from arbitrary
objects in the list of records, allowing us to use named tuples or
other objects that also hold additional information.

It's not entirely clear defining the generic iteration interface this
way is the best way to go, but it's convenient for the moment.

This currently iterates over all records, but we may want normally to
iterate just over CODE segment records, which we might do with this or
might do by forcing the client to use an interface that more clearly
states what it's doing.