testmc.symtab: Add sections to symbols
Sections (under various names) are a general concept in most
toolchains so we bring this up to the generic level. In ASxxxx these
are "areas" and in AS these are "segments"; we use "section" as the
generic name as that avoids confusion with memory "segments" as used
in the addressing of some processors and for other reasons described
in more detail in the comments.

There is no default value for the section when creating a Symbol
because the toolchain-specific code should be clear about whether it
supports sections or not.