testmc.asl: Macro Assembler AS .map file parsing and SymTab creation
This does not currently parse names for local scopes ("sections" in
asl terminology) but just leaves the section number in brackets
appended to the symbol (e.g., `foo[0]`). This could be used as-is, but
the number to name mapping can easily change (e.g., because a new
local scope was added) so it's useful in only with special care in the
most simple cases.

If any String symbol values have spaces in them this requires a
version of Macro Assembler AS with the fix to make it print them as
`\032` instead of ` ` in the map file. The bug exists from about 2008
through asl-current-142-bld151, but a version with the patch is
currently also available in the github.com/0cjs/asl repo.

If a space is embedded in the middle of a String symbol value this
will trigger a ParseError that will mention the possibility of the ASL
version being wrong, but leading and trailing spaces will be silently
dropped from values since we have no way to tell that they should be
there.