Test: Allow assembler directives as params to asbuild()
To have files that we can both include in "main" builds and also
separately unit test, we need to be able to leave the ORG directive
out of source code files for individual modules, letting the master
compile determine where the code will be replaced. Thus, we need a
way to get an ORG back in when assembling the unit test version.

We do this by allowing the specification of some initial assembler
directives with the `asbuild` command; since we always assemble only
one "master" file, these can be all additional parameters to the
`asbuild` command.

This lets us specify the origin to use for the unit test builds by
adding `ORG location` to those `asbuild` commands. We do this
initially just for `simple-asl.a65`.