src/std.a65: Extract ds/db/dw/zds macros
These we intend to use across much of our 6502 code, so we extract
them to a file that can be included into other files.

We do not protect against multiple includes, which will break the
assembly. It's not clear that we ever should do this; it seems better
at the moment to let the top-level master assembly file determine what
should be included for whatever particular modules it's pulling in.

The tests currently remain in simple-asl.a65 and simple-asl_test.py
for lack of a better place to put them right now. They should probably
be extracted at some point.