src/bigint.a65: Convert from ASxxxx to asl (Macro Assembler AS)
Some of the labels are awkward because though they're actually
ultra-local (e.g., skipping the increment of the MSB of a word when
the LSB doesn't roll over) we can't use `-` and `+` because these
"nameless" temporary symbols actually generate named symbols (with
unique names, but not starting with a dot) that then close and reopen
the scope of `.foo` (composed) temporary symbols and `$$foo` (named)
temporary symbols.

This is arguably a bug, or at least deficiency, in AS, but if it's
been around this long it's unlikely to be fixed in the near future.
For at least some of these cases, however, we may be able to get
around this with a macro, depending on how a macro implements its
"internal" symbols.