src/bigint_test: bi_read_dec buffer offset edge case test
This catches the case where we might not do a correct word increment
of the input pointer when moving past a leading sign character. We
didn't actually have this bug, but I noticed how easily it might
appear.

We actually test (in a limited way) across a range of addresses; we
are still missing complete coverage, however. For example, there's
still no case that fails if the word increment at the end of
.normalize is changed to a byte increment without carrying when
there's a rollover.