testmc.asl: Ignore blank lines before paragraphs in .map file
If no source was generated from the assembly, the `Segment CODE`
section will not be at the start of the file, but there will
still be an empty line before the `Symbols in Segment NOTHING`
paragraph. This produced a ParseError because that line did not
start with one of the paragraph heading lines we know.

To fix this we now allow and ignore any number of blank
(whitespace-only) lines before a paragraph.