src/heap heapinit: First try at setting up a heap
This is feeling pretty inefficient; we may need to store more of the
heap header information in the zero page. Or possibly, for simplicity,
we shouldn't divide the heap into a "used area with free areas in it"
and a separate "free area"; it just makes allocation slower (except
perhaps in some very specialized cases) without adding anything
helpful.