src/objects.md: Free space block objects (FBO) description
All free space on the heap is now in heapdata objects of the FBO type
because this makes implementation both significantly more efficient
and easier than using (nil,nil) cons cells. (The work that shows this
is currently on a development branch.)

We also now distinguish between the allocation size (asize) of
objects, which is always dword aligned, and the data size used for
heapdata objects (hdsize) which need not fill the entire allocated
space.

Also better describe pointers into the heap and fix some typos.