testmc.m6502.Registers: Ensure flags are always None, 0 or 1
It's not clear if this is required by the emulator, but we require
this in __repr__(). We could change repr to use `==` instead of `is`,
but we know that 0 and 1 work in the emulator so it's more reliable to
always use those than to assume the emulator can also handle, e.g.,
True and False instead.