src/biginit bi_x10: Remove an unneeded loop
Previously we first multiplied the input buffer by 2 (shifted left by
1) and then did a second loop to copy this to the scratch buffer. This
second loop isn't necessary; we can store the ×2 value in the scratch
buffer in the same loop as the first multiplication.

This reduces the code size from 55 to 48 bytes (a 7 byte or 13%
reduction) and speeds up the routine by about 11%:
    1 byte:  109→97 cycles (-11.01%)
    2 bytes: 200→178 cycles (-11.00%)
  255 bytes: 24741→21936 cycles (-11.34%)

Thanks to vtk on forum.6502.org[1] for inspiring this.

[1]: http://forum.6502.org/viewtopic.php?f=2&t=5835&p=72588#p72588