testmc.m6502: deposit functions now take multiple parameters
As well as giving a byte or list of bytes/words to deposit, such as
[1,2,3] or b'abc', we can now give multiple arguments. This allows
us to conveniently mix argument types, such as ints and bytestrings.

A typical use is to make it easier to deposit guard bytes without
having to ensure all data are the same type and build a list manually.
This:

    data = b'abcd'; M.deposit(addr, b'\xEE' + data + b'\xEF')

can now be:

    data = b'abcd'; M.deposit(addr, 0xEE, data, 0xEF)

We also now return the bytes deposited as a list of ints, thinking
that it might be useful in later comparisons. But that API may change
later.

The same has been done for word deposits. The return value is a list
of the words as Python integers.