src/objects: swap heapdata/char tag bits; routine to dispatch on ref type
We introduce a sample routine that shows how to determine the type of
and decode a reference (tagged pointer) to a heap cons cell, heapdata
header, char/byte or smallint.

In writing this we realized that life is a bit easier if we swap the
tag bits for heapdata (%01→%10) and char/byte (%10→%01), so update the
documentation for that as well.

We also make the following documentation changes:

* Note that the char/byte type currently ignores bits 7-2 of the LSB;
  we may be able to use this for something later.
* Change the valid type IDs from 1-63 to 0-63, and set the free block
  object type ID to 0 instead of 1. Trying to make 0 special in some
  way seems to serve no purpose and making things less general makes
  for slower code.
* Add a note that the current assignments for "informational" bits in
  the type IDs/formats (e.g., "is this object a number") may be
  sub-optimal since the BIT instruction gives us special treatment for
  bits 7 and 6. We'll want to look later at how these are assigned.