src/simple-asl: Change conflicting symbol introduced by AS bld152
In AS bld152, unnamed temporary symbols (defined with `+`, `-` and
`/`) no longer create a new scope for named temporary symbols starting
with `.`, so `.foo` now translates not to `__forw1.foo` but instead
`named8.foo`, the last used non-unnamed temporary symbol. Rename our
third `.foo` to `.foo2` to demonstrate this and avoid the name
collision with the second `.foo`.