bin/{a1send,wozmon-deposit}: Upload .p files to my Apple 1 replica
wozmon-deposit reads the records in a Macroassembler AS .p file
and prints out a list of Woz monitor deposit commands to deposit
the data into memory. It will append `.p` to the filename if the
filename has no periods in it, and if there are no slashes will
automatically find the file in `.build/obj/src/`. Errors cause an
exception to be thrown and backtrace to be printed; developers
who cannot understand backtraces have bigger problems than
dealing with errors from this program. :-)

a1send runs wozmon-deposit on the given files and sends the
deposit commands to /dev/ttyUSB0 using minicom's ascii-xfr
program, which is expected to be connected to an Apple 1 running
the Woz monitor.

a1send uses a 50 ms delay between characters and a 150 ms delay
between lines. This seems to be reasonably reliable, but let me
know if there are issues on your system, and feel free also to
try shorter delays and let me know how it works.

You currently need to reference the assembler listing to get the
addresses you want to call; I don't see any XXX

Potential improvements:
- Add an option to specify the tty to which to upload the data.
- Allow sending only specific address ranges from the .p file, to
  make updates of large programs faster.
- Add an option to read the symbol table of the given assembly
  (.map file) and generate code to JSR to a given symbol,
  followed by a JMP $FF1F to return to the monitor afterwards.
  It could send the R command to run that bit of code, too.