bin/wozmon-deposit: Add -c/-n argument for \r or \n newlines
We want to send only \r characters to the monitor because the \n chars,
though harmless, cause line feeds in the output we see. (That's a bug
in the RC6502 replica's screen→serial code.)

That didn't work, however, because, though we can use `ascii-xfr -n`
to pass our \r sequences straight through without translation,
ascii-xfr inserts a line delay only when it sees \n, not \r, and so
overflows the Apple 1 input routine.

The fix is probably to have wozmon-deposit send to the serial port
itself, waiting to see an echo of each char it sends before sending
the next. (This would also ensure we're sending at near the maximum
possible speed.) However, that's a lot more complex because to read
from the serial line it needs to stop the terminal program (probably
with SIGSTOP) to avoid the terminal program eating characters that
wozmon-deposit needs to read.