tools/asl/: Rework for updated build system for Macro Assembler AS
As of asl-current-142-bld156 the Makefile.defs has changed to add
cross-build functionality, which changes some variable names. We now
also have a new Makefile.def-x86_64-unknown-linux which is actually
multi-architecture (it has nothing x86_64 specific in it and just
builds for the current architecture), so we use that instead of or old
custom native-unknown-linux defs file.