src/objects: Change format to allow for mark bit if using separate heaps
We change the tag format so that bit 0 clear is used for fixed
contstants, pointers and smallints, and bit 0 set indicates a obdata
header. (We needed to remove the char/unsigned byte type to do this.)

Now if we use separate heaps for cons cells and obdata, when GC'ing a
heap we know that type from the heap and no longer need bit 0 to
distingush them. This frees up bit 0 to use (with care) as a mark bit
while GC'ing a heap.

Split heaps may speed allocation and GC, too, just by virtue of the
cons cell heap having one fixed size of object.