conftest.py: Run .pt files as tests and fix filename collision problem
This pytest plugin collects *.pt files as tests. As an added bonus,
this uses new (Python ≥ 3.5) importlib machinery to avoid adding the
test module names to sys.modules, avoiding module name collisions for
tests with the same name in different directories. It also no longer
needs to add the test file's directories to sys.path, removing yet
another source or error.

*_test.py and test_*.py are still collected in the normal way and
still suffer from the collision problem. (Pytest as of 5.4.1 still has
no standard solution for this. LocalPath.pyimport() in py 1.8.1 does
support an ensuresyspath='importlib' option, designed for use by
pytest, that appears as if it would fix the problem, but there seems
to be no way to configure pytest to use it.)