tool/asl/Setup: Replace Build with somewhat different Python script
We want to extend this build system to other tools as well, and Bash
was getting too complex for this.

The behaviour of the new script is probably slightly different, but
neither the old nor the new behaviour is exactly what we want yet. In
particular, we need to work out how to deal with wanting to use a
local build when one is already provided by the environment. (Perhaps
a flag or environment variable to force it?)