bin/{ascii2a2t,t2a2dsk}: Convert/copy ASCII file to Apple II disk image
ascii2a2t reads an ASCII file and prints it to stdout in Apple II
format, converting NL to CR and setting the high bit. It dies if it
reads a non-ASCII character.

t2a2dsk takes a disk image and a list of ASCII files, converts each
one to Apple II text format using ascii2a2t, and writes it to the
given disk image under the same basename as the input file.