testmc.memimage: Get memory image data as contiguous bytes
contigbin() returns the data in the memory image as a contiguous
sequence of bytes from the lowest to the highest address, raising an
error if there is any overlap, and contiglen() returns the length of
this sequence.

These can be used to easily generate a binary file from a MemImage.