Test a2exebuild(): Build disk image with executable program
This copies the empty DOS 3.3 diskette image we pulled in via
tool/diskimg/, writes our Apple II binary program to it, and writes a
new HELLO program that will BRUN that binary. It also sets the
auto-run program name to HELLO, just in case.

Currently a comment gives the command to run the LinApple emulator on
the diskette image, but we should add an option to the Test program to
do this for us.