exe/a2/mondump: Load and hook src/a2/mondump hexdump command
We prepend a the code to hook the hexdump command to the Monitor's
Ctrl-Y "user" command JMP location. This code is at the end of the
input buffer, which is fine because it can be wiped out after it's run
once by the BRUN command.

To allow the executable framework to choose the location of the
hexdump code, we remove the `org` from src/a2/mondump, instead adding
`org`s to exe/a2/mondump.a65 and the build command for the unit test
version. (The latter is in a different location just to help ensure
that hexdump is really relocatable, though that's not actually tested
properly.)

We rely on p2a2bin's check for overlapping records to ensure that the
setup code doesn't overwrite any of the actual hexdump code.