a2/mondump → moncmds, and move dumpmonvar here from misc
When not explicitly specified, hexdump uses default parameters from
the Monitor's saved address locations. While working out how to to
deal better with this, it turns out to be very handy to have
dumpmonvar available to dump them.

This brings dumpmonvar into the moncmds executable as the first step
towards turning the Ctrl-Y routine into one that parses multiple
commands. Currently the functionality is unchanged, however. (One can
call `300G` to run dumpmonvar, but the `300` will overwrite the
locations it's dumping.)