src/a2/moncmds dumpmonvar: Document and fix incorrect newline printing
It's supposed to print a newline _before_ printing its output, not after.

It was incorrectly using CROUT1 at the end, rather than CROUT,
meaning that output would go to the current output device but the
newline would be printed only on the screen.

Also, update a comment in nextline now that we better understand
why PRA1 is doing what we need.