src/a2/moncmds usrcmd: Implement commands under Ctrl-Y
Ctrl-Y followed by 'A' now calls dumpmonvar, and plain Ctrl-Y (i.e.,
followed by anything else, which may be the CR terminating the input
line or further other chars for the monitor to parse) has been changed
to call hexdump.

This is slightly tricky because the monitor continues parsing the
command line after we're done and there must be at least one more
command: the trailing CR is parsed by the monitor as "stop parsing"
when it's not the first character on the input line. See the comments
for more details on this.