Test: Unless -A given once, do not build or test ASxxxx code
ASxxxx is used only for legacy code, and requires special support from
the OS to run 32-bit binaries. To make life easier for people trying
out this repo, we now by default do not install ASxxxx, use it to
assemble files, or run the tests for those files. However, passing the
-A flag to Test just once will permanently enable ASxxxx and all its
tools.

The switch that enables this is the existence of $basedir/.all-tools
(created by specifying the -A option). This not under the .build/
directory so that it persists across fully clean (-C) rebuilds.

This is a bit hacky at the moment in two ways, but still better than
nothing:
1. It adds more things requiring the current working directory to be
   $basedir when running the tests.
2. We really should have a more generic way of specifying which sets
   of tools are wanted, especially the DOS tools, emulators, etc.

The README is also still very out of date on this and use of the
current default assembler, AS, and needs to be extensively rewritten.