src/simple-{asl,asx}.pt: Extract ident symbol/string check to separate test
There's no need to check this on every load; if it works on the first
call, it will work the same on subsequent ones.

This is also not a check we can reasonably pull up into a shared `M`
fixture because there's no generic information in the symbol table
(for either ASL or ASxxxx) containing information about the filename.
It's not worth requiring that for every file because the mere symbol
names will provide plenty of indication about what's wrong, most of
the time. (It might be useful to print or store some information about
what file was loaded, however.)