tool/linapple/: Hack for broken dependencies broken and no longer needed
The res/*.xpm object files have been moved, so those targets no longer
exist, breaking our hack that builds them first. However, that part of
the dependency graph appears to have been fixed, so we no longer need
the hack anyway.

But other parts of the dependency graph may still be broken; the
current workaround if it fails due to that is just to run the build
again.

We do leave in a comment noting this for people trying to build
an older version.