tool/linapple/: New path for linapple binary
This gets this repo working again with linapple's master branch head.

Obviously it's now going to break with older versions, but we'll just
let developers who want to get this to build an older version deal
with that themselves. (For most developers, a better option is just to
build the older version separately and make sure that its `linapple`
binary is in the path when doing a rebuild here; that will be used and
this linapple won't be built.)