mc6800sim: Implement step() and the LDAA immediate opcode
This now starts to outline the general structure of the opcode interpreter.
We have a mapping of opcodes to functions that take a machine state and
update it appropriately for the execution of that opcode. We also map
opcodes to unique mnemonic names so that we can create an Instructions
class that does the reverse mapping, making it easy to generate machine
langauge in tests.

Currently step() will throw a NotImplementedError if the opcode it's trying
to execute is not in the OPCODES map. This can go away once the OPCODES map
is completed; even illegal opcodes should have a mapping that does the
right thing (probably blows up the interpreter with an exception, since
this interpreter is primarily aimed at unit testing machine code).

The one test we have tests both the step() function and the LDAA immediate
opcode in detail. Possibly this needs to be split up. It's not clear how
we'll move forward on this. The level of testing we need for fairly
complete coverage of a full opcode interpreter is fairly enormous; it would
be nice to have some sort of external testing suite like the many for the
6502.[1]

[1]: http://visual6502.org/wiki/index.php?title=6502TestPrograms