mc6800sim.memory: No longer provide memory storage, just access methods
This will allow it to be used with other simulators that provide their own
memory storage.

We do this by making the class abstract and forcing the class hosting this
mixin to provide a get_memory_seq() function that returns a reference to a
mutable sequence representing the memory.

The tests now need to be expanded to test different underlying storage
(e.g., a list of Integer as used by py65) and that deposit() etc. are not
allowing values that won't fit in a byte to be stored even if the
underlying storage permits it.

We also need to make this configurable for big- or little-endian word
access.