mc6800sim.memory: Document backing store and test with [int]
Update the class to mention that get_memory_seq() needs to be overridden,
and the get_memory_seq() docstring to explain what kind of backing store we
need,

We parametrize the tests so all are run with both the original bytearray()
backing store and now a list of int, `[0]*65536`. This necessitated
removing the tests that the backing store was not permitting invalid byte
values to be inserted, but we should no longer rely on this now that the
deposit functions have been confirmed to check the value before depositing.