mc6800sim.memory: byte() and bytes() always return int and bytes
bytes() already always converted its result to a `bytes`, but we now more
clearly test this.

byte() did not ensure its return value is an `int`, though the tests never
failed because all our test backing stores always did. There could possibly
be a a sequence holding bytes that doesn't return an int when indexing a
single value, but it's hard to imagine what it might be or look like, so I
can think of no simple test for this. Nonetheless, we simply pass the value
to the int() constructor before returning it, which is fine and should
catch any violations of our backing store contract.