testmc.m6502: Convert to use testmc.memory
The replaces old examine/deposit/etc. code in m6502.Machine with the
testmc.memory.MemoryAccess mixin. There are still a few functions there
that should be moved to the new MemoryAccess class, such as the string
examine function. (Though perhaps now that bytes() returns a bytes object
which has string-like literals we don't really need it any more?)

Rather than just deleting the old tests we convert them. This helps to
ensure that the replacement worked correctly, but having done this the
tests can be deleted after checking for any coverage the m6502 tests have
that the testmc.memory ones don't.

The main interface change is that M.bytes() now returns a `bytes` object
rather than a list of `int`; this required updating some of the tests under
src/. This demonstrates that writing such tests becomes a bit cleaner now
because the bytestrings are textually shorter than the lists, especially
when using hexadecimal instead of decimal.