testmc.registers: set_attrs_on()
This allows us to set the non-None register and flag values as attributes
of the same name on another object, a "machine," giving us a generic
interface for setting register and flag values on any machine that has a
Registers class defined for it. (Getting the current register/flag values
is done merely by passing those attributes to a Registers() constructor.)

For flags it can either directly set boolean flag attributes or change bits
in a status register attribute.