testmc.registers: Status register is now optional
Normally a machine would have either a status register or an individual
bool attribute for each flag; we make the status register optional to
support the latter. If `srname` is not defined (or is defined to Null)
we set the individual booleans only.

Since without a status register set_attrs_on(o, setsr=True) cannot work, we
add a better exception message for the AttributeError that raises. However
setting the wrong attributes (sr when it should be setting individual flag
attributes or vice versa) should never happen to normal end users since the
machine should always have a setregs() that uses it properly.

All that said, it's not clear if perhaps we shouldn't be offering a way to
synthesize a status register from the flags and `srbits` definition, since
we have that information there. Technically that's the job of the machine
that chose not to use flag attributes rather than a status register, but it
might be nice to offer generic code to help out such machines.