testmc.machine: GenericMachine implementing regs(), setregs(), etc.
This defines (and partially implements) the interface for a "machine" class
whose instances simulate a microprocessor system with memory and registers.
The sublasses may be the simulator class itself or a wrapper around a
simulator that adapts the interface.

If the machine has a `regsobj` attribute, we store the register and flags
or status register values as attributes on the object referenced by
`regsobj`. Otherwise register and flag/SR values are stored as attributes
on the machine object itself.

This does not yet implement or force the subclass to implement any part
of the step()/call()/etc. interface; that's probably the next step.