testmc.m6502: Inherit from GenericMachine now
The main change here is that we move our Registers class to an
attribute of the Machine and sublcass GenericRegisters.

As usual, the register-related tests on m6502.Machine are updated for this
rather than removed; this helps confirm that behaviour hasn't changed. In
addition, though we remove the old m6502.Register class, we keep and
convert its tests for the same reason. All these can be removed later,
though it's probably a good idea to keep a smoke test or two for the use of
registers on m6502.Machine.

To src/conftest.py we add a new R() fixture that returns the Registers
attribute of the Machine object created by the M() fixture, the same as we
already do for the symtab attribute in the S() fixture, and update all the
tests to use the new Registers class.